add Accept: header handling, and support JSON, XML, and plain text

+++ api/src/main/java/com/redhat/gss/victims/Check.java
@@ -7,6 +7,7 @@ import javax.servlet.http.Part;
 import javax.ws.rs.GET;
 import javax.ws.rs.POST;
 import javax.ws.rs.Consumes;
+import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
@@ -14,6 +15,7 @@ import javax.ws.rs.core.Context;
 
 import com.redhat.chrometwo.api.security.SecurityInterceptor;
 import javax.mail.internet.ContentDisposition;
+import javax.mail.internet.ParseException;
 
 import java.io.ByteArrayInputStream;
 
@@ -42,11 +44,14 @@ import java.util.Map;
 import org.jboss.resteasy.plugins.providers.multipart.InputPart;
 import org.jboss.resteasy.plugins.providers.multipart.MultipartFormDataInput;
 
+import javax.xml.bind.annotation.XmlRootElement;
+import javax.xml.bind.annotation.XmlElement;
+import javax.xml.bind.annotation.XmlAttribute;
+
 @Path("/check")
 @Stateless
 @LocalBean
 public class Check {
-
 
     public Check() {
 
@@ -62,42 +67,61 @@ public class Check {
         }
     }
 
+    @POST
+	@Consumes({ MediaType.MULTIPART_FORM_DATA })
+    @Produces({ MediaType.TEXT_PLAIN, "text/*", "*/*" })
+    public String checkString(MultipartFormDataInput inputForm, @Context HttpServletRequest request) throws VictimsException, IOException, ParseException  {
+
+        List<CheckResultElement> checkResult = check(inputForm, request);
+        StringBuilder result = new StringBuilder();
+        for (CheckResultElement checkResultElement : checkResult) {
+            String checkFileName = checkResultElement.getFile();
+            List<String> cves = checkResultElement.getVulnerabilities();
+            if (cves != null && cves.size() > 0) {
+                result.append(String.format("%s VULNERABLE! ", checkFileName));
+                for (String cve : cves) {
+                    result.append(cve);
+                    result.append(" ");
+                }
+                result.append("\n");
+            } else {
+                result.append(checkFileName + " ok\n");
+            }
+        }
+
+        return result.toString();
+    }
 
     @POST
 	@Consumes({ MediaType.MULTIPART_FORM_DATA })
-    public String checkMulti(MultipartFormDataInput inputForm, @Context HttpServletRequest request) throws Exception {
+    @Produces({ MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML, MediaType.TEXT_XML })
+    @org.jboss.resteasy.annotations.providers.jaxb.Wrapped(element="checkresult")
+    public List<CheckResultElement> check(MultipartFormDataInput inputForm, @Context HttpServletRequest request) throws VictimsException, IOException, ParseException {
 
-        StringBuilder result = new StringBuilder();
         StringBuilder trace = new StringBuilder();
         
         trace.append("multi: ");
         trace.append(displayHeaders(request));
 
-        VictimsDBInterface db;
-        VictimsResultCache cache;
-
-        try {
-            db = VictimsDB.db();
-            cache = new VictimsResultCache();
-
-        } catch (VictimsException e) {
-            result.append("VictimsException while opening the database:\n");
-            e.printStackTrace();
-            return result.append(e.toString()).append("\n").toString();
+        boolean foundTraceMarker = false;
+        String[] values = request.getParameterValues("trace");
+        if (values != null) {
+            foundTraceMarker = true;
+            for (String value : values ) {
+                trace.append("trace value: " + value + "\n");
+            }
+        } else {
+            trace.append("trace value: null\n");
         }
 
-        try {
-            trace.append("About to synchronize local database with upstream ...\n");
-            db.synchronize();
-            trace.append("   successful synchronize.\n");
+        VictimsDBInterface db = VictimsDB.db();
+        VictimsResultCache cache = new VictimsResultCache();
 
-        } catch (VictimsException e) {
-            result.append("VictimsException while synchronize-ing local database:\n");
-            e.printStackTrace();
-            return result.append(e.toString()).append("\n").toString();
-        }
+        trace.append("About to synchronize local database with upstream ...\n");
+        db.synchronize();
+        trace.append("   successful synchronize.\n");
 
-
+        List<CheckResultElement> checkResult = new ArrayList<CheckResultElement>();
         boolean foundAtLeastOne = false;
         Map<String, List<InputPart>> multiValuedMap = inputForm.getFormDataMap();
         for (Map.Entry<String, List<InputPart>> entry : multiValuedMap.entrySet()) {
@@ -121,19 +145,42 @@ public class Check {
 
                 ContentDisposition disp = new ContentDisposition(dispString);
                 String fileName = disp.getParameter("filename");
-                if (fileName == null) {
-                    fileName = name;
+                CheckResultElement checkResultElement = null;
+
+                // this is only used for debugging return values from victims
+                if (name.equals("victimsdebug") && fileName == null) {
+                    String s = inputPart.getBodyAsString();
+                    trace.append("victimsdebug: " + s);
+                    String[] a = inputPart.getBodyAsString().split("\\s+");
+                    trace.append(" victimsdebug: " + a);
+                    if (a.length > 0) {
+                        checkResultElement = new CheckResultElement();
+                        checkResultElement.setFile(a[0]);
+                        trace.append(" victimsdebugfile: " + a[0]);
+                        for (int i = 1; i < a.length; i++) {
+                            checkResultElement.addVulnerability(a[i]);
+                            trace.append(" victimsdebugvuln: " + a[1]);
+                        }
+                    }
+                } else {
+                    if (fileName == null) {
+                        fileName = name;
+                    }
+
+                    String tmpFileName = null;
+                    try {
+                        tmpFileName = copyToTempFile(fileName, inputPart.getBody(InputStream.class, null));
+                        checkResultElement = checkOne(db, cache, tmpFileName);
+
+                    } finally {
+                        if (tmpFileName != null) {
+                            deleteTempFile(tmpFileName);
+                        }
+                    }
                 }
 
-                String tmpFileName = null;
-                try {
-                    tmpFileName = copyToTempFile(fileName, inputPart.getBody(InputStream.class, null));
-                    result.append(checkOne(db, cache, tmpFileName));
-
-                } finally {
-                    if (tmpFileName != null) {
-                        deleteTempFile(tmpFileName);
-                    }
+                if (checkResultElement != null) {
+                    checkResult.add(checkResultElement);
                 }
             }
         }
@@ -142,88 +189,74 @@ public class Check {
             trace.append("no parts found\n");
         }
         trace.append("end of results\n");
-        return result.toString();
+
+        return checkResult;
     }
 
-    private String checkOne(VictimsDBInterface db, VictimsResultCache cache, String tmpFileName) throws Exception {
+    private CheckResultElement checkOne(VictimsDBInterface db, VictimsResultCache cache, String tmpFileName) throws VictimsException, IOException {
         // tmpFileName is the full (absolute) file name of the temporary copy of the uploaded file
         // it's last path element should be the name of the file that the user attached to it in the uploaded request
  
-            String fileName = new File(tmpFileName).getName();
+        String fileName = new File(tmpFileName).getName();
 
-            StringBuilder result = new StringBuilder();
-            StringBuilder trace = new StringBuilder();
+        StringBuilder trace = new StringBuilder();
        
-            trace.append("filename: ").append(tmpFileName).append("\n");
+        trace.append("filename: ").append(tmpFileName).append("\n");
 
-            // the cache was giving incorrect results (saying things were OK when they were not, 
-            // or vise-versa.  It may be behaving this way because i was screwing with the local 
-            // database alot while i was getting this working, or there may be some deeper problem.
-            // for now just turn off the cache.
-            String key = null;   //checksum(tmpFileName);
-            trace.append("key: ");
-            trace.append(key);
-            trace.append("\n");
+        // the cache was giving incorrect results (saying things were OK when they were not, 
+        // or vise-versa.  It may be behaving this way because i was screwing with the local 
+        // database alot while i was getting this working, or there may be some deeper problem.
+        // for now just turn off the cache.
+        String key = null;   //checksum(tmpFileName);
+        trace.append("key: ");
+        trace.append(key);
+        trace.append("\n");
             
-            // Check cache 
-            if (key != null && cache.exists(key)) {
-                try {
-                    HashSet<String> cves = cache.get(key);
-                    if (cves != null && cves.size() > 0) {
-                        result.append(String.format("%s VULNERABLE! ", fileName));
-                        for (String cve : cves) {
-                            result.append(cve);
-                            result.append(" ");
-                        }
-                        result.append("\n");
-                    } else {
-                        result.append(fileName + " ok\n");
-                    }
-                } catch (VictimsException e) {
-                    result.append("VictimsException while checking cache:\n");
-                    e.printStackTrace();
-                    return result.append(e.toString()).append("\n").toString();
+        // Check cache 
+        if (key != null && cache.exists(key)) {
+            CheckResultElement checkResultElement = new CheckResultElement();
+            checkResultElement.setFile(fileName);
+            HashSet<String> cves = cache.get(key);
+            if (cves != null && cves.size() > 0) {
+                trace.append(String.format("%s VULNERABLE! ", fileName));
+                for (String cve : cves) {
+                    checkResultElement.addVulnerability(cve);
+                    trace.append(cve);
+                    trace.append(" ");
                 }
+                trace.append("\n");
+            } else {
+                trace.append(fileName + " ok\n");
             }
+            return checkResultElement;
+        }
 
-            // Scan the item
-            ArrayList<VictimsRecord> records = new ArrayList();
-            try {
-
-                VictimsScanner.scan(tmpFileName, records);
-                for (VictimsRecord record : records) {
-
-                    try {
-                        HashSet<String> cves = db.getVulnerabilities(record);
-                        if (key != null) {
-                            cache.add(key, cves);
-                        }
-                        if (!cves.isEmpty()) {
-                            result.append(String.format("%s VULNERABLE! ", fileName));
-                            for (String cve : cves) {
-                                result.append(cve);
-                                result.append(" ");
-                            }
-                            result.append("\n");
-                        } else {
-                            result.append(fileName + " ok\n");
-                        }
-
-                    } catch (VictimsException e) {
-                        result.append("VictimsException while checking database:\n");
-                        e.printStackTrace();
-                        return result.append(e.toString()).append("\n").toString();
-                    }
+        // Scan the item
+        ArrayList<VictimsRecord> records = new ArrayList();
+        VictimsScanner.scan(tmpFileName, records);
+        CheckResultElement checkResultElement = new CheckResultElement();
+        checkResultElement.setFile(fileName);
+        for (VictimsRecord record : records) {
+            HashSet<String> cves = db.getVulnerabilities(record);
+            if (key != null) {
+                cache.add(key, cves);
+            }
+            if (!cves.isEmpty()) {
+                trace.append(String.format("%s VULNERABLE! ", fileName));
+                for (String cve : cves) {
+                    checkResultElement.addVulnerability(cve);
+                    trace.append(cve);
+                    trace.append(" ");
                 }
-            } catch (IOException e) {
-                result.append("VictimsException while scanning file:\n");
-                e.printStackTrace();
-                return result.append(e.toString()).append("\n").toString();
+                trace.append("\n");
+            } else {
+                trace.append(fileName + " ok\n");
             }
-            return result.toString();
+        }
+        return checkResultElement;
     }
 
-    private String displayHeaders(HttpServletRequest request) throws Exception {
+    private String displayHeaders(HttpServletRequest request) {
         StringBuilder result = new StringBuilder();
         for (java.util.Enumeration<java.lang.String> headerNames = request.getHeaderNames();
              headerNames.hasMoreElements();) {
@@ -257,6 +290,34 @@ public class Check {
 
         return hash;
     }
+
+    @XmlRootElement
+    static public class CheckResultElement {
+        private String file;
+        private List<String> vulnerabilities = new ArrayList<String>();
+
+        @XmlAttribute
+        public String getFile() {
+            return file;
+        }
+
+        public void setFile(String file) {
+            this.file = file;
+        }
+
+
+        @XmlElement(name="vulnerability")
+        public List<String> getVulnerabilities() {
+            return vulnerabilities;
+        }
+
+        public void addVulnerability(String vulnerability) {
+            if (vulnerability != null && !vulnerability.isEmpty()) {
+                this.vulnerabilities.add(vulnerability);
+            }
+        }
+    }
+
 
     // 
     // Both the cache and the local database want/need to read the whole file seperately, so we must save
